Linked_Learning/pricing_sample/fastapi_app/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
pydantic==2.7.4
python-multipart==0.0.9
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
python-dateutil==2.9.0
